+++ src/main/java/com/market/member/application/auth/AuthService.java
@@ -3,8 +3,8 @@ package com.market.member.application.auth;
 import com.market.global.event.Events;
 import com.market.member.application.auth.dto.LoginRequest;
 import com.market.member.application.auth.dto.SignupRequest;
-import com.market.member.domain.auth.RegisteredEvent;
 import com.market.member.domain.auth.TokenProvider;
+import com.market.member.domain.auth.event.RegisteredEvent;
 import com.market.member.domain.member.Member;
 import com.market.member.domain.member.MemberRepository;
 import com.market.member.domain.member.NicknameGenerator;
